+++ src/settings.js
@@ -11,7 +11,7 @@ export const registerSettings = function() {
         name: "MaterialDeck.Sett.Enable",
         scope: "global",
         config: true,
-        default: false,
+        default: true,
         type: Boolean,
         onChange: x => window.location.reload()
     });
@@ -50,6 +50,13 @@ export const registerSettings = function() {
         
     });
 
+    //Create the Help button
+    game.settings.registerMenu(MODULE.moduleName, 'helpMenu',{
+        name: "MaterialDeck.Sett.Help",
+        label: "MaterialDeck.Sett.Help",
+        type: helpMenu,
+        restricted: true
+    });
     /**
      * Playlist soundboard
      */
@@ -110,3 +117,45 @@ export const registerSettings = function() {
         restricted: true
     });
 }
+
+export class helpMenu extends FormApplication {
+    constructor(data, options) {
+        super(data, options);
+    }
+  
+    /**
+     * Default Options for this FormApplication
+     */
+    static get defaultOptions() {
+        return mergeObject(super.defaultOptions, {
+            id: "helpMenu",
+            title: "Material Deck: "+game.i18n.localize("MaterialDeck.Sett.Help"),
+            template: "./modules/MaterialDeck/templates/helpMenu.html",
+            width: "500px"
+        });
+    }
+  
+    /**
+     * Provide data to the template
+     */
+    getData() {
+      
+        return {
+           
+        } 
+    }
+  
+    /**
+     * Update on form submit
+     * @param {*} event 
+     * @param {*} formData 
+     */
+    async _updateObject(event, formData) {
+  
+    }
+  
+    activateListeners(html) {
+        super.activateListeners(html);
+        
+    }
+  }